requirements.txt
black
flake8
isort
pip
python-dotenv
-e .
bar==0.2.1
barbar==0.2.1
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
filelock==3.14.0
fonttools==4.53.0
fsspec==2024.5.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-ml-py3==7.352.0
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pretty-errors==1.2.25
psutil==5.9.8
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.5.0
scikit-video==1.1.11
scipy==1.13.1
six==1.16.0
sympy==1.12.1
threadpoolctl==3.5.0
torch==2.2.2
torch-utils==0.1.2
torchsummary==1.5.1
torchvision==0.17.2
tqdm==4.66.4
typing_extensions==4.12.1
tzdata==2024.1

